pyproject.toml
[project]
name = "calendar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "playwright>=1.58.0",
    "playwright-stealth>=2.0.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "ruff>=0.15.0",
]
